requirements.txt
streamlit
requests
google-generativeai
python-dotenv
pandas